Rename base package as common
