+++ app/src/main/java/com/yesferal/hornsapp/app/presentation/common/BaseFragment.kt
@@ -1,4 +1,4 @@
-package com.yesferal.hornsapp.app.presentation.base
+package com.yesferal.hornsapp.app.presentation.common
 
 import android.os.Bundle
 import android.view.View
